Add a t_force function, which force a lazy object into a normal object.
This function recursively calls all the 'SELECT id ...' of the sub-objects, in order to populate the object with the values from the database. In order to deal with recursive data, this function takes a cache and only do a SELECT if the pair (table name, id) is not in the table. So from now, all the t_get* functions first build a lazy object, and then force or not the object depending of the values of _lazy, given as argument to the get function.

Last thing: the force function is not really optimized at the moment, I think it should call a bit too often SELECT. As always, this can be improve later :-)