Check that the database we are connecting to contains valid types.
At the moment, we only check that the types we are defining are valid subtypes of the one in the database. It would be nice to do the opposite checks when we do a save call, or to define RO and RW database where we check subtyping or type compatibilty respectively.

I think we can use Types.t to refactore some bits of code in pa_orm as it can leads to easily extensible code (ie. just need to define a function type_of_mytype: mytype -> Types.t to be able to persist values of type mytype. That can be usefull for external piece of code where we do not control the type definition and thus cannot add the 'with ...' constraint.