Try to handle external types which are not mutually recursive.
The goal is to handle something like:
type x = Foo | Bar with orm ()
type y = { foo: x } with orm ()

It was almost working, but it seems that I hit a weird type-conv bug here (see lib_test/bug.ml). Need to report that to Markus to see what he thinks it is.