Add a new libray value-of. It aims to do the same kind of things as type-of, but for values instead.
The goal is to have a nice finite representation of any infinite values (using some fix-point representation, as Rec x.Y where x is bounded to Y inside Y). In the opposite of types, it is not possible to build that representation at pre-processing time, as we do not know type associated to values. However, it is possible to build some functions t_of_value and value_of_t which will unroll the value cleverly to obtain what we want.

What I push is not complete yet, ie. it does not work for value with foreign fields (as the unrolling thing is not implemented yet). The unrolling thing is already done in type-of in pre-processing time, so it seems not too difficult to use it to generate the right function. Will do it in the next couple of days.

The next step will be to use write a small library which convert that value into SQL rows and do the correct select/insert/update/delete.

The schema creation is handled using type-of and seems to work fine on my limited testing.

So hopefully, in few days we will have a sql_backend completly decoupled from the type/value intermediate representation generator. Which is something quite nice, I believe :-) (as json/xmlrpc backend already exists for this kind of intermediate values).