Lot of work done on the value library - it's almost working and it is quite elegant.
The idea is to generate an intermediate structure more simple then usual ML values, but which will have the same sharing, ie. if a ML value got some link to an other ML value, the corresponding intermediate structure will have the same link.

When the recursion problem will be fixed (very soon hopefully) I will start working on the SQL backend, which will take as input the intermediate types/values.

And also, I think the decomposition works for almost all the ML types, which is also quite neet.