[value] Make the conversion Value.t -> ML works.
It is using Obj.magic for cyclic values, as I couldn't see an another way to do it properly. Moreover, cyclic values are allowed only if defined via record fields. I believe it should be possible to make them work for variant types, but that would mean playing more deeply with the purple magic of the Obj module and I don't really like messing with that ...