[lib] modify the way we are storing lists and arrays, to be more consistent with the new immutable database schema.
Now lists are stored as cons cells, with the size until the end. We use big JOIN statements to get and check if the list has already been saved, which is maybe not very safe on very big lists.

The overhall algo is there, but it doens't work correctly at the moment ...