Use the simpler weakid library in ORM.
Instead of generating a huge record for storing weak IDs for each types, use a single one for the Obj.t type. In the current state, it may not be very safe but the short-term plan is to rewrite that part to use a global weakid table for each pair (type, database) to be able to have modify mutable objects in the database.