[lib] clean-up the deletion semantics;
Now, when we delete a value, it is deleted from the database only if there are not other value referencing it in the database; to be sure that a value is deleted, you can try a t_id on it just after, if it works, that means that there are other objects depending on it.

This also works for the sub-values of the deleted value: if nothing else that the current value is referecing them, there are deleted from the database; otherwise they are still there.

The way I plan to handle recursive values is to delete everything which is in the recursive cycle only if nothing is referenced outside; should be not too tricky to do this. That aslo means that you will have to manually unlink two values if you want to delete one and not the other (which seems safer to me).

Last thing, I got an unexpected 'segmentation fault' when I do an SQL query inside a SQL custom function; it seems to come from the binding (sql_init.ml:65 is triggering it) I haven't look at it yet - that also means that I havent tested that code too much at the moment :-)